This commit was manufactured by cvs2git to create tag 'MIRBSD_10_BASE'.

+++ src/lib/libc/stdio/fgets.c
@@ -75,7 +75,7 @@ fgets(char *buf, int n, FILE *fp)
 		 * newline, and stop.  Otherwise, copy entire chunk
 		 * and loop.
 		 */
-		if (len > n)
+		if (len > (size_t)n)
 			len = n;
 		t = memchr((void *)p, '\n', len);
 		if (t != NULL) {
