+++ src/sys/dev/pcmcia/com_pcmcia.c
@@ -1,3 +1,4 @@
+/**	$MirOS: src/sys/dev/pcmcia/com_pcmcia.c,v 1.2 2005/03/06 21:27:54 tg Exp $ */
 /*	$OpenBSD: com_pcmcia.c,v 1.39 2005/01/27 17:04:55 millert Exp $	*/
 /*	$NetBSD: com_pcmcia.c,v 1.15 1998/08/22 17:47:58 msaitoh Exp $	*/
 
@@ -319,7 +320,7 @@ found:
 		return;
 	}
 
-	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, psc->sc_pcioh.size);
+	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, (int)(psc->sc_pcioh.size));
 
 	sc->sc_iobase = -1;
 	sc->enable = com_pcmcia_enable;
@@ -340,7 +341,7 @@ found:
 
 #ifdef notyet
 	sc->enabled = 0;
-	
+
 	com_pcmcia_disable1(sc);
 #endif
 }
@@ -429,7 +430,7 @@ com_pcmcia_disable1(sc)
 	pcmcia_function_disable(psc->sc_pf);
 }
 
-/* 
+/*
  * XXX This should be handled by a generic attach
  */
 void
@@ -506,7 +507,7 @@ com_pcmcia_attach2(sc)
 	}
 #endif
 #endif
-	
+
 	/* Reset the LCR (latch access is probably enabled). */
 	bus_space_write_1(iot, ioh, com_lcr, lcr);
 	if (sc->sc_uarttype == COM_UART_16450) { /* Probe for 8250 */
